# Conflicts:
#	db/changelog/changelog-3.5.xml
#	service/src/test/resources/com/aplana/sbrf/taxaccounting/script/service/RefBookServiceImplTest.xml

+++ model/src/main/java/com/aplana/sbrf/taxaccounting/model/ndfl/NdflPersonIncome.java
@@ -84,7 +84,7 @@ public class NdflPersonIncome extends NdflPersonOperation {
     protected String paymentNumber;
 
     // НДФЛ.Перечисление в бюджет.Платежное поручение.Сумма (Графа 24)
-    protected Long taxSumm;
+    protected BigDecimal taxSumm;
 
     // Дата и время редактирования. Заполняется при редактировании данных НФ через загрузку Excel файла
     protected Date modifiedDate;
